multiboxloss conf_data debugging for tpu

+++ layers/modules/multibox_loss.py
@@ -147,7 +147,7 @@ class MultiBoxLoss(nn.Module):
 #             pos_idx = pos_idx.to(device)
 #             neg_idx = neg_idx.to(device)
             
-        both_idx = (pos_idx + neg_idx).gt(0)
+        both_idx = (pos_idx + neg_idx).gt(0).long()
         conf_p = conf_data[both_idx]
         conf_p = conf_p.view(-1, self.num_classes)
 #         conf_p = conf_p[(pos_idx+neg_idx).gt(0)].view(-1, self.num_classes)
